pyproject.toml
[project]
name = "nyaya-ams"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "pypdf>=6.7.4",
    "pytest>=9.0.2",
    "sentence-transformers>=5.2.3",
]
